requirements.txt
celery~=5.0
Django~=3.1
djangorestframework~=3.12
django-celery-beat~=2.1
django-celery-results~=2.0
django-environ~=0.4
django-extensions~=3.1
django-filter~=2.4
django-model-utils~=4.0
django-rest-auth~=0.9
numpy~=1.18
psycopg2-binary~=2.8